remove spec versioning scaffold for now; add missing doc comments
